% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRDprediction.R
\name{HRDprediction}
\alias{HRDprediction}
\title{HRD Prediction}
\usage{
HRDprediction(data)
}
\arguments{
\item{data}{From function signimercopy}
}
\value{
HRDCNA probability scores of the samples.
}
\description{
a robust HRD predictor "HRDCNA" (Homologous Recombination
Deficiency Prediction by Copy Number Feature) based on CNA features.
}
\examples{
score_swgs <- HRDprediction(data = nmfcn_swgs)
score_wgs <- HRDprediction(data = nmfcn_wgs)
score_snp <- HRDprediction(data = nmfcn_snp)

}
