% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigminercopy.R
\name{sigminercopy}
\alias{sigminercopy}
\title{Calling Copy Number Features}
\usage{
sigminercopy(data, genome_build = c("hg19", "hg38", "mm10", "mm9"))
}
\arguments{
\item{data}{Absolute copy number profile}

\item{genome_build}{Genome build version, should be 'hg19', 'hg38', 'mm9' or 'mm10'.}
}
\value{
The matrix of copy number features
}
\description{
Detail information described in Sigminer.
}
\examples{
cn_swgs <- readRDS("./data/test/cn_60_wgs10x.rds")
nmfcn_swgs <- sigminercopy(data = cn_swgs, "hg19")

cn_wgs <- readRDS("./data/test/cn_60_wgs.rds")
nmfcn_wgs <- sigminercopy(data = cn_wgs, "hg19")

cn_snp <- readRDS("./data/test/cn_60_snp.rds")
nmfcn_snp <- sigminercopy(data = cn_snp, "hg19")

}
